src/EcclesiaCRM/utils/MiscUtils.php : deltree bug resolution php8

+++ src/EcclesiaCRM/utils/MiscUtils.php
@@ -352,9 +352,11 @@ class MiscUtils {
  * @param string $dir the directory name
  */
   public static function delTree($dir) {
-   $files = array_diff(scandir($dir), array('.','..'));
+    if ( ! is_dir($dir) ) return false;
+
+    $files = array_diff(scandir($dir), array('.','..'));
     foreach ($files as $file) {
-      (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
+        (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
     }
     return rmdir($dir);
   }
